requirements.txt
librosa==0.10.1
scipy==1.10.1
noisereduce==3.0.0
pyannote.audio==3.0.0
faster-whisper==0.9.0
pydub==0.25.1
torch==2.0.1
numpy==1.24.3
pydantic==2.10.4
python-dotenv==1.0.1